+++ test/e2e/btc_rewards_distribution_e2e_test.go
@@ -867,7 +867,7 @@ func AddCovdSigsToPendingBtcDels(
 		pendingDel, err := chain.ParseRespBTCDelToBTCDel(pendingDelResp)
 		require.NoError(t, err)
 
-		if pendingDel.HasCovenantQuorums(bsParams.CovenantQuorum) {
+		if pendingDel.HasCovenantQuorums(bsParams.CovenantQuorum, 0) {
 			continue
 		}